Revert "recover"
This reverts commit 751c0aabd294502dc366f59ca9b09072cffbbbb8.

+++ video.go
@@ -17,7 +17,7 @@ func AddVideoTag(c *gin.Context) {
 	uVid := int(vVid)
 	strTagId := c.PostForm("tagid")
 	vTagId, _ := strconv.Atoi(strTagId)
-	uTagId := int16(vTagId)
+	uTagId := int(vTagId)
 	DBaddVideoTag(uVid, uTagId)
 }
 
@@ -142,7 +142,7 @@ func DBaddVideoComment(vid int, author int, text string) {
 	return
 }
 
-func DBaddVideoTag(vid int, tagid int16) {
+func DBaddVideoTag(vid int, tagid int) {
 	tag := Tag{Tid: tagid, Kind: 1, Pid: vid}
 	db.Insert(tag)
 	return
